refactor: :recycle: Move code structure to nextjs style
