+++ assets/TechnologyIcons.json
