+++ components/HomePage/ProjectHighlights.tsx
