+++ components/HomePage/TechnicalSkills.tsx
