+++ package.json
@@ -1,69 +1,29 @@
 {
-  "name": "e-portfolio",
-  "version": "0.1.0",
-  "private": true,
-  "homepage": "https://pumbas.net",
-  "dependencies": {
-    "@emotion/react": "^11.7.0",
-    "@emotion/styled": "^11.6.0",
-    "@mui/icons-material": "^5.2.0",
-    "@mui/material": "^5.2.1",
-    "@testing-library/jest-dom": "^5.15.0",
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/user-event": "^12.8.3",
-    "@types/jest": "^26.0.24",
-    "@types/node": "^12.20.37",
-    "@types/react": "^17.0.34",
-    "@types/react-dom": "^17.0.11",
-    "eslint-config-next": "^13.0.2",
-    "next": "^13.0.2",
-    "node-sass": "^7.0.1",
-    "postcss-flexbugs-fixes": "^5.0.2",
-    "postcss-preset-env": "^7.4.3",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-router-dom": "^6.0.2",
-    "react-transition-group": "^4.4.2",
-    "set-value": "^4.1.0",
-    "typescript": "^4.4.4",
-    "web-vitals": "^1.1.2"
-  },
-  "scripts": {
-    "start": "react-scripts start",
-    "predeploy": "react-scripts build",
-    "deploy": "gh-pages -d build",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  },
-  "devDependencies": {
-    "autoprefixer": "^10.4.13",
-    "eslint-config-airbnb": "^19.0.0",
-    "eslint-plugin-import": "^2.25.3",
-    "eslint-plugin-jsx-a11y": "^6.5.1",
-    "eslint-plugin-react": "^7.27.0",
-    "eslint-plugin-react-hooks": "^4.3.0",
-    "gh-pages": "^3.2.3",
-    "postcss": "^8.4.18",
-    "react-scripts": "^5.0.1",
-    "tailwindcss": "^3.2.1"
-  }
+	"name": "e-portfolio",
+	"version": "0.1.0",
+	"private": true,
+	"homepage": "https://pumbas.net",
+	"dependencies": {
+		"next": "^13.0.2",
+		"node-sass": "^7.0.1",
+		"react": "18.2.0",
+		"react-dom": "18.2.0"
+	},
+	"scripts": {
+		"dev": "next dev",
+		"build": "next build",
+		"start": "next start",
+		"lint": "next lint"
+	},
+	"devDependencies": {
+		"@types/node": "18.11.9",
+		"@types/react": "18.0.25",
+		"@types/react-dom": "18.0.8",
+		"autoprefixer": "^10.4.13",
+		"eslint": "8.26.0",
+		"eslint-config-next": "^13.0.2",
+		"postcss": "^8.4.18",
+		"tailwindcss": "^3.2.1",
+		"typescript": "^4.8.4"
+	}
 }
