--- src/Components/Utils/LinkBase.tsx
@@ -1,22 +1,0 @@
-import React from "react";
-
-interface LinkBaseProps {
-    href: string;
-    internal?: boolean;
-    className?: string;
-}
-
-const LinkBase: React.FC<LinkBaseProps> = (props) => {
-    return (
-        <a target={props.internal ? undefined : '_blank'}
-           rel="noreferrer"
-           href={props.href}
-           className={`${props.className ?? ''} inline-block`}
-        >
-            {props.children}
-        </a>
-    )
-}
-
-export default LinkBase;
-export type { LinkBaseProps };
